requirements.txt
streamlit
pandas
openpyxl
python-docx
PyPDF2
pdfplumber
pytesseract
Pillow
kagglehub
